Add CANStructs for MotorControllerCAN

+++ Common/lib/CANStructs/ECUCANStructs.h
@@ -32,7 +32,7 @@ public:
         (motor_on, 1)
     )
 
-    uint16_t get_message_ID()
+    uint32_t get_message_ID()
     {
         return ECUMotorCommands_MESSAGE_ID;
     }
@@ -65,7 +65,7 @@ public:
         (battery_contact, 1)
     )
 
-    uint16_t get_message_ID()
+    uint32_t get_message_ID()
     {
         return ECUPowerAuxCommands_MESSAGE_ID;
     }
